requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
